fix small issues

+++ l8/database/migrations/2021_01_15_125653_add_created_by_user_id_to_products_table.php
@@ -14,7 +14,7 @@ class AddCreatedByUserIdToProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->unsignedInteger('created_by_user_id');
+            $table->unsignedBigInteger('created_by_user_id');
             $table->foreign('created_by_user_id')->references('id')->on('users');
         });
     }
